CI: Skip updating npm on Windows
The Microsoft hosted agents are updated frequently to the latest
versions of e.g. Node.js and npm.

"Updating" to a specific version ends up lagging behind, without
receiving bugfix updates. We can consider pinning again
if an npm 7.x is released, and we have to debug and vet that update.